readtags: introduce printtags.[ch] and move functions for printing tags to the files
